+++ src/main/java/uz/smartup/academy/bloggingplatform/service/LikeService.java
@@ -7,6 +7,6 @@ public interface LikeService {
     LikeDTO findByUserAndPost(int userId, int postId);
     void addLike(int userId, int postId);
     void removeLike(int userId, int postId);
-    int countLikesByPostId(int postId);
+    long countLikesByPostId(int postId);
 
 }
